IMPALA-12471: Add unit tests of external jdbc tables for MySQL
This patch adds MySql tests for the "external data source"
mechanism in Impala to implement data source for querying JDBC.

This patch also fixes the handling of case-sensitive table and
column names for MySQL query.

Testing:
- Added unit test for mysql and ran unit-test with JDBC
driver mysql-connector-j-8.1.0.jar. This test requires
to add the docker to sudoer's group. Also, the test is
only run in 'exhaustive' mode.

Change-Id: I446ec3d4ebaf53c8edac0b2d181514bde587dfae